requirements.txt
openai
opencv-python
python-dotenv